backend/requirements.txt
# MYC3LIUM Backend Requirements

# Core API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6

# Data processing
numpy>=1.24.0
scipy>=1.11.0

# Serial/Hardware (optional - install on Raspberry Pi)
pyserial>=3.5
spidev>=3.6; platform_machine == "armv7l" or platform_machine == "aarch64"
smbus2>=0.4.2; platform_machine == "armv7l" or platform_machine == "aarch64"

# Meshtastic (optional - install on device with Heltec V3)
meshtastic>=2.0.0
pypubsub>=4.0.3

# Security
cryptography>=41.0.0

# Authentication
PyJWT>=2.8.0
argon2-cffi>=23.1.0

# Configuration
tomli_w>=1.0.0

# Environment
python-dotenv>=1.0.0

# Logging
python-json-logger>=2.0.0
pydantic-settings>=2.0.0
